Function intersectLine() created
.intersectLine( line : Line3, target : Line3 ) : Boolean
line - test the intersection between this and line.
target - the result will be copied into this Line3. This line connects the closest points between the two lines (ie. it is the only line that is perpendicular on both lines).

Returns true if the intersection lies within start and end on both lines.